+++ src/views/Streets/Streets.js
@@ -1,25 +1,27 @@
-import React, { Component, Fragment } from 'react';
-import { connect } from 'react-redux';
-import { Link } from 'react-router-dom';
-import classnames from 'classnames';
-import { ApplicationLayout } from '../../layouts';
-import { TextInput, Spinner } from '../../components';
-import { getIncidences } from '../../redux/actions/states';
-import { getAreas, sortList } from '../../redux/actions/areas';
+import React, { Component, Fragment } from "react";
+import { connect } from "react-redux";
+import { Link } from "react-router-dom";
+import classnames from "classnames";
+import swal from "sweetalert";
+import { ApplicationLayout } from "../../layouts";
+import { TextInput, Spinner } from "../../components";
+import { getIncidences } from "../../redux/actions/states";
+import { getAreas, sortList } from "../../redux/actions/areas";
 import {
     getStreets,
     updateStreet,
     deleteStreet,
     addStreet
-} from '../../redux/actions/streets';
+} from "../../redux/actions/streets";
 
 class Lgas extends Component {
     state = {
         areas: [],
         streets: [],
         active: {},
-        name: '',
-        areaId: '',
+        name: "",
+        sort: "",
+        areaId: "",
         report: {
             kidnap: 0,
             murder: 0,
@@ -27,11 +29,11 @@ class Lgas extends Component {
         },
         selected: null,
         view: null,
-        order: ''
+        order: ""
     };
 
     componentDidMount() {
-        const { getAreas, getStreets, getIncidences, incidences } = this.props;
+        const { getAreas, getStreets, getIncidences } = this.props;
         getAreas();
         getStreets();
         getIncidences();
@@ -47,7 +49,7 @@ class Lgas extends Component {
         const report = JSON.parse(street.report);
         this.setState({
             selected: street,
-            view: '',
+            view: "",
             name: street.name,
             report
         });
@@ -71,11 +73,11 @@ class Lgas extends Component {
     };
 
     createView = () => {
-        this.setState({ selected: null, view: 'add', name: '', areaId: '' });
+        this.setState({ selected: null, view: "add", name: "", areaId: "" });
     };
 
     removeItem = () => {
-        this.setState({ selected: null, view: '', name: '', areaId: '' });
+        this.setState({ selected: null, view: "", name: "", areaId: "" });
     };
 
     handleDropdown(item) {
@@ -123,19 +125,34 @@ class Lgas extends Component {
     deleteItem = () => {
         const { deleteStreet } = this.props;
         const { selected } = this.state;
-        deleteStreet(selected.id);
-        this.setState({ selected: null });
+        swal({
+            title: "Are you sure?",
+            text: "Once deleted, you will not be able to undo action!",
+            icon: "warning",
+            buttons: true,
+            dangerMode: true
+        }).then(willDelete => {
+            if (willDelete) {
+                deleteStreet(selected.id);
+                this.setState({ selected: null });
+                swal("Poof! item has been deleted!", {
+                    icon: "success"
+                });
+            } else {
+                return;
+            }
+        });
     };
 
     sortData = () => {
-        const { order } = this.state;
+        const { order, sort } = this.state;
         const { sortList } = this.props;
-        if (!order || order === 'desc') {
-            this.setState({ order: 'asc' });
-            sortList('asc');
+        if (!order || order === "desc") {
+            this.setState({ order: "asc" });
+            sortList("asc", sort);
         } else {
-            this.setState({ order: 'desc' });
-            sortList('desc');
+            this.setState({ order: "desc" });
+            sortList("desc", sort);
         }
     };
 
@@ -146,6 +163,7 @@ class Lgas extends Component {
             view,
             name,
             areaId,
+            sort,
             report,
             errors
         } = this.state;
@@ -155,14 +173,21 @@ class Lgas extends Component {
             <ApplicationLayout>
                 <div className="hidden lg:flex flex-col relative z-10 w-full max-w-xs flex-grow flex-shrink-0 border-r bg-gray-300">
                     <div className="flex-shrink-0 px-4 py-2 flex items-center justify-between border-b bg-gray-200">
-                        <button className="flex items-center text-xs font-semibold text-gray-600">
-                            Sorted by Name
-                            <svg
-                                viewBox="0 0 24 24"
-                                className="ml-1 h-6 w-6 fill-current text-gray-500">
-                                <path d="M7.293 9.293a1 1 0 011.414 0L12 12.586l3.293-3.293a1 1 0 111.414 1.414l-4 4a1 1 0 01-1.414 0l-4-4a1 1 0 010-1.414z"></path>
-                            </svg>
-                        </button>
+                        <select
+                            className={classnames(
+                                "form-select outline-none mt-1 block text-sm bg-gray-200 text-gray-600",
+                                {
+                                    "border-red-500": false,
+                                    className: true
+                                }
+                            )}
+                            placeholder="Sort by ..."
+                            name="sort"
+                            onChange={this.handleChange}
+                            value={sort}>
+                            <option value="">Sort by Name</option>
+                            <option value="rating">Sort by Ratings</option>
+                        </select>
                         <button onClick={this.sortData}>
                             <svg
                                 viewBox="0 0 24 24"
@@ -250,9 +275,9 @@ class Lgas extends Component {
                                                                                             item
                                                                                         )}
                                                                                         className={classnames(
-                                                                                            'inline-flex items-center text-sm outline-none',
+                                                                                            "inline-flex items-center text-sm outline-none",
                                                                                             {
-                                                                                                'font-bold':
+                                                                                                "font-bold":
                                                                                                     !selected ||
                                                                                                     selected.id !==
                                                                                                         item.id
@@ -428,7 +453,7 @@ class Lgas extends Component {
 
                         {!view && selected && (
                             <div className="flex items-center">
-                                <div style={{ width: '60%' }}>
+                                <div style={{ width: "60%" }}>
                                     <h2>{selected.name}</h2>
 
                                     <div className="mt-2">
@@ -446,6 +471,7 @@ class Lgas extends Component {
                                                     }
                                                     placeholder="Enter State Name"
                                                     label="Name"
+                                                    error={errors.name}
                                                 />
                                             </div>
                                             <div className="w-full">
@@ -460,7 +486,7 @@ class Lgas extends Component {
                                                     }
                                                     placeholder="Select Area"
                                                     label="Area"
-                                                    error={null}
+                                                    error={errors.areaId}
                                                 />
                                             </div>
                                             <div className="w-full text-gray-700 my-2">
@@ -544,7 +570,7 @@ class Lgas extends Component {
                                         </form>
                                     </div>
                                 </div>
-                                <div className="ml-20" style={{ width: '20%' }}>
+                                <div className="ml-20" style={{ width: "20%" }}>
                                     <div className="p-4 border border-gray-300 rounded shadow">
                                         <h4 className="text-base leading-tight uppercase text-center font-medium">
                                             Sicura Rating
@@ -557,8 +583,8 @@ class Lgas extends Component {
                             </div>
                         )}
 
-                        {view === 'add' && (
-                            <div style={{ width: '60%' }}>
+                        {view === "add" && (
+                            <div style={{ width: "60%" }}>
                                 <h2 className="uppercase mb-4 text-gray-800 font-bold">
                                     Add Street
                                 </h2>
@@ -573,7 +599,7 @@ class Lgas extends Component {
                                                 handleChange={this.handleChange}
                                                 placeholder="Enter Street Name"
                                                 label="Name"
-                                                error={null}
+                                                error={errors.name}
                                             />
                                         </div>
                                         <div className="w-full">
@@ -586,7 +612,7 @@ class Lgas extends Component {
                                                 handleChange={this.handleChange}
                                                 placeholder="Select Area"
                                                 label="Area"
-                                                error={null}
+                                                error={errors.areaId}
                                             />
                                         </div>
                                         <div className="w-full text-gray-700 my-2">
